feat: update logic and workflow

+++ cinema-be/src/controllers/movieCategory.controller.js
@@ -1,5 +1,6 @@
 const movieCategoryRepository = require('../repositories/movieCategory.repository');
 const nextId = require('../utils/nextId');
+const { assertMovieOwnership } = require('../utils/movieOwnership');
 
 async function list(req, res) {
   const mappings = await movieCategoryRepository.findAll();
@@ -16,6 +17,9 @@ async function create(req, res) {
   if (movie_id === undefined || cat_id === undefined) {
     return res.status(400).json({ message: 'movie_id and cat_id are required' });
   }
+  if (!(await assertMovieOwnership(req, movie_id))) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
 
   const id = await nextId('movieCategory');
   const mapping = await movieCategoryRepository.create({ id, movie_id, cat_id });
@@ -23,6 +27,10 @@ async function create(req, res) {
 }
 
 async function removeForMovie(req, res) {
+  if (!(await assertMovieOwnership(req, req.params.movieId))) {
+    return res.status(403).json({ message: 'Forbidden' });
+  }
+
   await movieCategoryRepository.deleteByMovieId(req.params.movieId);
   res.json({ message: 'Deleted' });
 }
